[be] refactor: response package refactor
