+++ server/src/main/java/com/CreatorConnect/server/auth/handler/MemberAuthenticationEntryPoint.java
@@ -1,6 +1,6 @@
 package com.CreatorConnect.server.auth.handler;
 
-import com.CreatorConnect.server.errorresponse.ErrorResponder;
+import com.CreatorConnect.server.response.ErrorResponder;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.AuthenticationException;
