+++ server/src/main/java/com/CreatorConnect/server/auth/handler/MemberAuthenticationFailureHandler.java
@@ -1,6 +1,6 @@
 package com.CreatorConnect.server.auth.handler;
 
-import com.CreatorConnect.server.errorresponse.ErrorResponse;
+import com.CreatorConnect.server.response.ErrorResponse;
 import com.google.gson.Gson;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
